+++ include/util/hashtable.h
@@ -12,10 +12,10 @@ typedef struct {
 typedef struct {
 	unsigned int (*hash)(const void *);
 	hashtable_entry_t **buckets;
-	int bucket_count;
+	size_t bucket_count;
 } hashtable_t;
 
-hashtable_t *create_hashtable(int buckets, unsigned int (*hash_function)(const void *));
+hashtable_t *create_hashtable(size_t buckets, unsigned int (*hash_function)(const void *));
 void free_hashtable(hashtable_t *table);
 void *hashtable_get(hashtable_t *table, const void *key);
 void *hashtable_set(hashtable_t *table, const void *key, void *value);
