+++ include/util/list.h
@@ -1,9 +1,11 @@
 #ifndef _LIST_H
 #define _LIST_H
 
+#include <stddef.h>
+
 typedef struct {
-	int capacity;
-	int length;
+	size_t capacity;
+	size_t length;
 	void **items;
 } list_t;
 
@@ -11,14 +13,16 @@ list_t *create_list(void);
 void list_free(list_t *list);
 void list_foreach(list_t *list, void (*callback)(void* item));
 void list_add(list_t *list, void *item);
-void list_insert(list_t *list, int index, void *item);
-void list_del(list_t *list, int index);
+void list_insert(list_t *list, size_t index, void *item);
+void list_del(list_t *list, size_t index);
 void list_cat(list_t *list, list_t *source);
 // See qsort. Remember to use *_qsort functions as compare functions,
 // because they dereference the left and right arguments first!
 void list_qsort(list_t *list, int compare(const void *left, const void *right));
 // Return index for first item in list that returns 0 for given compare
 // function or -1 if none matches.
-int list_seq_find(list_t *list, int compare(const void *item, const void *cmp_to), const void *cmp_to);
+int list_seq_find(list_t *list,
+		int compare(const void *item, const void *cmp_to),
+		const void *cmp_to);
 
 #endif
