+++ src/util/list.c
@@ -4,6 +4,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stddef.h>
 
 #include "util/list.h"
 
@@ -34,7 +35,7 @@ void list_foreach(list_t *list, void (*callback)(void *item)) {
 	if (list == NULL || callback == NULL) {
 		return;
 	}
-	for (int i = 0; i < list->length; i++) {
+	for (size_t i = 0; i < list->length; i++) {
 		callback(list->items[i]);
 	}
 }
@@ -44,20 +45,20 @@ void list_add(list_t *list, void *item) {
 	list->items[list->length++] = item;
 }
 
-void list_insert(list_t *list, int index, void *item) {
+void list_insert(list_t *list, size_t index, void *item) {
 	list_resize(list);
 	memmove(&list->items[index + 1], &list->items[index], sizeof(void*) * (list->length - index));
 	list->length++;
 	list->items[index] = item;
 }
 
-void list_del(list_t *list, int index) {
+void list_del(list_t *list, size_t index) {
 	list->length--;
 	memmove(&list->items[index], &list->items[index + 1], sizeof(void*) * (list->length - index));
 }
 
 void list_cat(list_t *list, list_t *source) {
-	int i;
+	size_t i;
 	for (i = 0; i < source->length; ++i) {
 		list_add(list, source->items[i]);
 	}
@@ -67,8 +68,10 @@ void list_qsort(list_t* list, int compare(const void *left, const void *right)) 
 	qsort(list->items, list->length, sizeof(void *), compare);
 }
 
-int list_seq_find(list_t *list, int compare(const void *item, const void *data), const void *data) {
-	for (int i = 0; i < list->length; i++) {
+int list_seq_find(list_t *list,
+		int compare(const void *item, const void *data),
+		const void *data) {
+	for (size_t i = 0; i < list->length; i++) {
 		void *item = list->items[i];
 		if (compare(item, data) == 0) {
 			return i;
